[fix] ci settings and related files

+++ bin/arch_install.sh
@@ -1,8 +1,12 @@
-sudo pacman -Syu
-mkdir /tmp/yay
-cd /tmp/yay
-curl -OJ 'https://aur.archlinux.org/cgit/aur.git/plain/PKGBUILD?h=yay'
-makepkg -si
-cd
-rm -rf /tmp/yay
-yay --version
+#!/bin/bash
+
+if [[ $(uname -r) =~ ARCH$ ]]; then
+    sudo pacman -Syu --noconfirm
+    mkdir /tmp/yay
+    cd /tmp/yay || exit
+    curl -OJ 'https://aur.archlinux.org/cgit/aur.git/plain/PKGBUILD?h=yay'
+    makepkg -si --noconfirm
+    cd || exit
+    rm -rf /tmp/yay
+    yay --version
+fi
